+++ R/data.R
@@ -42,3 +42,25 @@ NULL
 #'
 #'
 "liposomes_227"
+
+#' dat_1.
+#'
+#' FLUOstar .dat files used for creation of the update and unusable for data immediate data analysis.
+#'
+#'
+"dat_1"
+
+#' dat_2.
+#'
+#' FLUOstar .dat files used for creation of the update and unusable for data immediate data analysis.
+#'
+#'
+"dat_2"
+
+#' dat_3.
+#'
+#' FLUOstar .dat files used for creation of the update and unusable for data immediate data analysis. 
+#' This file is unique because it validates a major bug fix to ensure that users get the right output.
+#'
+"dat_3"
+